[Console] Generate url components with the new script
