+++ src/plugins/console/server/lib/spec_definitions/json/overrides/nodes.usage.json
@@ -1,0 +1,10 @@
+{
+  "nodes.usage": {
+    "url_components": {
+      "metrics": [
+        "_all",
+        "rest_actions"
+      ]
+    }
+  }
+}
